Group authentication code into auth package
